Updated deferred shading shader and implemented D3DLocalInclude
The code is in place to count the number of lights in a tile's frustum.
D3DLocalInclude is used to override the default behavior of the HLSL
compiler's include so that a specific root directory can be designated
for all shaders. Texture changed to Texture2D and includes new
initialization utilities. Also added overrides for the default copy
constructor and equal operator on classes that should not be copyable.